Add columns with date calculations based on refernce date

Add columns with date calculations based on refernce date

add_date_columns(
  df,
  date_col,
  ref_date = Sys.Date(),
  fiscal_year_offset = 6,
  week_start = 1,
  drop = FALSE
)

Arguments

df
data frame

date_col
column with class of 'date'

ref_date
reference date for calculations, defaults to current date

fiscal_year_offset
the number of months to offset date, if fiscal year ends in June, use 6

week_start
integer for start of week where Monday = 1 and Sunday = 7

drop
some of the generated fields may match the input data frame. When TRUE, the original columns will be removed and replaced with the new field of the same name. Otherwise, columns with the same name will be appended with a '1'

Examples

demo_data() %>%
  add_date_columns(date_col = date)

#> # A tibble: 10 x 11
#>    group     x     y     z date         day  week month quarter calendar_year
#>    <chr> <dbl> <dbl> <dbl> <date>     <dbl> <dbl> <dbl>   <dbl>         <dbl>
#>  1 a       101    10     1 2021-03-04     0     0     0       0             0
#>  2 a       102    20     0 2021-01-04   -59    -8    -2       0             0
#>  3 b       103    30     1 2020-11-04  -120   -17    -4      -1            -1
#>  4 b       104    40     0 2020-09-04  -181   -26    -6      -2            -1
#>  5 c       105    50     1 2020-07-04  -243   -35    -8      -2            -1
#>  6 c       106    60     0 2020-05-04  -304   -43   -10      -3            -1
#>  7 d       107    70     1 2020-03-04  -365   -52   -12      -4            -1
#>  8 d       108    80     0 2020-01-04  -425   -61   -14      -4            -1
#>  9 e       109    90     1 2019-11-04  -486   -69   -16      -5            -2
#> 10 e       110   100     0 2019-09-04  -547   -78   -18      -6            -2
#> # ... with 1 more variable: fiscal_year <dbl>