Add column of headlines

Add column of headlines

add_headline_column(
  df,
  x,
  y,
  headline = "{trend} of {delta} ({orig_values})",
  ...,
  .name = "headline",
  if_match = "There was no difference.",
  trend_phrases = headliner::trend_terms(),
  plural_phrases = NULL,
  orig_values = "{x} vs. {y}",
  n_decimal = 1,
  round_all = TRUE,
  multiplier = 1,
  return_cols = NULL
)

Arguments

df

data frame, must be a single row

x

a numeric value to compare to the reference value of 'y'

y

a numeric value to act as a control for the 'x' value

headline

a string to format the final output. Uses glue syntax

...

arguments passed to glue_data

.name

string value for the name of the new column to create

if_match

string to display if numbers match, uses glue syntax

trend_phrases

list of values to use for when y is more than x, y is the same as x, or y is less than x. You can pass it just trend_terms (the default) and call the result with "...{trend}..." or pass is a named list (see examples)

plural_phrases

named list of values to use when difference (delta) is singular (delta = 1) or plural (delta != 1)

orig_values

a string using glue syntax. example: ({x} vs {y})

n_decimal

numeric value to limit the number of decimal places in the returned values.

round_all

logical value to indicate if all values should be rounded. When FALSE, the values will return with no modification. When TRUE (default) all values will be round to the length specified by 'n_decimal'.

multiplier

number indicating the scaling factor. When multiplier = 1 (default), 0.25 will return 0.25. When multiplier = 100, 0.25 will return 25.

return_cols

arguments that can be passed to select, ex: c("a", "b"), starts_with,etc.

Examples

# You can use 'add_headline_column()' instead of
# `mutate(headline = map2_chr(...))`
# here is an example comparing the sleeping habits of animals
head(animal_sleep) %>%
  dplyr::select(common_name, hours_asleep, hours_awake) %>%
  add_headline_column(
    x = hours_asleep,
    y = hours_awake,
    headline = "The {common_name} spends {delta} more {hours} {trend} than not {trend}.",
    trend_phrases = trend_terms(more = "asleep", less = "awake"),
    plural_phrases = list(hours = plural_phrasing(single = "hour", multi = "hours"))
  ) %>%
  knitr::kable("pandoc")
#> 
#> 
#> common_name                  hours_asleep   hours_awake  headline                                                                  
#> --------------------------  -------------  ------------  --------------------------------------------------------------------------
#> African Elephant                      3.3          20.7  The African Elephant spends 17.4 more hours awake than not awake.         
#> African Giant Pouched Rat             8.3          15.7  The African Giant Pouched Rat spends 7.4 more hours awake than not awake. 
#> Arctic Fox                           12.5          11.5  The Arctic Fox spends 1 more hour asleep than not asleep.                 
#> Arctic Ground Squirrel               16.6           7.4  The Arctic Ground Squirrel spends 9.2 more hours asleep than not asleep.  
#> Asian Elephant                        3.9          20.1  The Asian Elephant spends 16.2 more hours awake than not awake.           
#> Baboon                                9.4          14.6  The Baboon spends 5.2 more hours awake than not awake.                    

# you can also use 'return_cols' to return any and all "talking points".
# You can use tidyselect helpers like 'starts_with("delta")' or
# 'everything()'. In this example, I returned the delta & trend columns
# and identified the rows at the extremes
head(animal_sleep) %>%
  dplyr::select(common_name, hours_asleep, hours_awake) %>%
  add_headline_column(
    x = hours_asleep,
    y = hours_awake,
    headline = "more time {trend} ({orig_values} hours)",
    trend_phrases = trend_terms(more = "alseep", less = "awake"),
    return_cols = c("delta", "trend")
  ) %>%
  dplyr::filter(delta %in% range(delta)) %>%
  knitr::kable("pandoc")
#> 
#> 
#> common_name         hours_asleep   hours_awake  headline                                  delta  trend  
#> -----------------  -------------  ------------  ---------------------------------------  ------  -------
#> African Elephant             3.3          20.7  more time awake (3.3 vs. 20.7 hours)       17.4  awake  
#> Arctic Fox                  12.5          11.5  more time alseep (12.5 vs. 11.5 hours)      1.0  alseep