Adjust which words start with "a" vs "an"

headliner uses crude logic to anticipate the articles used before phrases like "(an) increase", "(a) decrease", "(an) 83" and "(a) -5" You will likely not call the function get_article() directly but you can augment the logic it uses with this function.

augment_article_patterns(regex_for_a, regex_for_an)

Arguments

regex_for_a

regular expression for words that get the article 'a'

regex_for_an

regular expression for words that get the article 'an'

Details

These arguments assume that the words start with the patterns you provide. Passing "hour|heir" will be translated to "^(hour|heir)".

See also

Examples

# The crude logic for headliner would give all "h" words the article "a"
# by default, the word "heirloom" returns "a"
get_article("heirloom")
#> Error in get_article("heirloom"): could not find function "get_article"

# the patterns that are used can be updated
augment_article_patterns(
  regex_for_a = "Euro|uni",
  regex_for_an = "hour|heir"
)

get_article("heirloom")
#> Error in get_article("heirloom"): could not find function "get_article"

# these patterns can be reset using 'NULL'
augment_article_patterns(
  regex_for_an = NULL
)

get_article("heirloom")
#> Error in get_article("heirloom"): could not find function "get_article"