For a list

For a list

For a data frame

# S3 method for list
headline(x, compare, reference, ...)

# S3 method for data.frame
headline(x, compare, reference, ...)

Arguments

x
data frame, must be a single row

compare
numeric value to compare against reference (base) value

reference
numeric value that 'compare' value will be compared against

...
Arguments passed on to compare_values

trend_phrasing

list of values to use for when y is more than x, y is the same as x, or y is less than x.

orig_values

a string using glue syntax. {c} = the 'compare' value, and {r} = 'reference'

n_decimal

numeric value to limit the number of decimal places in the returned values.

round_all

logical value to indicate if all values should be rounded. When FALSE, the values will return with no modification. When TRUE (default) all values will be round to the length specified by 'n_decimal'.

scale

number indicating the scaling factor. When scale = 1, 1/4 will return 0.25, when scale = 100 (default) 1/4 will return 25

Functions

headline.list: for lists

headline.data.frame: for data frames