Add columns with date calculations based on refernce date

Add columns with date calculations based on refernce date

add_date_columns(
  df,
  date_col,
  ref_date = Sys.Date(),
  fiscal_year_offset = 6,
  week_start = 1,
  drop = FALSE
)

Arguments

df
data frame

date_col
column with class of 'date'

ref_date
reference date for calculations, defaults to current date

fiscal_year_offset
the number of months to offset date, if fiscal year ends in June, use 6

week_start
integer for start of week where Monday = 1 and Sunday = 7

drop
some of the generated fields may match the input data frame. When TRUE, the original columns will be removed and replaced with the new field of the same name. Otherwise, columns with the same name will be appended with a '1'

Examples

# 0 indicates "current"
data.frame(date = seq(Sys.Date(), Sys.Date() - 100, by = "-20 days")) %>%
add_date_columns(
  date_col = date,
  ref_date = Sys.Date()
)

#>         date  day week month quarter calendar_year fiscal_year
#> 1 2020-09-05    0    0     0       0             0           0
#> 2 2020-08-16  -20   -3    -1       0             0           0
#> 3 2020-07-27  -40   -5    -2       0             0           0
#> 4 2020-07-07  -60   -8    -2       0             0           0
#> 5 2020-06-17  -80  -11    -3      -1             0          -1
#> 6 2020-05-28 -100  -14    -4      -1             0          -1