Add columns with date calculations based on reference date

Add columns with date calculations based on reference date

add_date_columns(
  df,
  date_col,
  ref_date = Sys.Date(),
  fiscal_year_offset = 6,
  week_start = 1,
  drop = FALSE
)

Arguments

df

data frame

date_col

column with class of 'date'

ref_date

reference date for calculations, defaults to current date

fiscal_year_offset

the number of months to offset date, if fiscal year ends in June, use 6

week_start

integer for start of week where Monday = 1 and Sunday = 7

drop

some of the generated fields may match the input data frame. When TRUE, the original columns will be removed and replaced with the new field of the same name. Otherwise, columns with the same name will be appended with a '1'

Examples

demo_data() %>%
  add_date_columns(date_col = date)
#> # A tibble: 10 × 11
#>    group sales count on_sale date         day  week month quarter calendar_year
#>    <chr> <dbl> <dbl>   <dbl> <date>     <dbl> <dbl> <dbl>   <dbl>         <dbl>
#>  1 a       101    35       1 2022-05-24     0     0     0       0             0
#>  2 a       102    34       0 2022-03-24   -61    -9    -2      -1             0
#>  3 b       103    33       1 2022-01-24  -120   -17    -4      -1             0
#>  4 b       104    32       0 2021-11-24  -181   -26    -6      -2            -1
#>  5 c       105    31       1 2021-09-24  -242   -35    -8      -3            -1
#>  6 c       106    30       0 2021-07-24  -304   -44   -10      -3            -1
#>  7 d       107    29       1 2021-05-24  -365   -52   -12      -4            -1
#>  8 d       108    28       0 2021-03-24  -426   -61   -14      -5            -1
#>  9 e       109    27       1 2021-01-24  -485   -70   -16      -5            -1
#> 10 e       110    26       0 2020-11-24  -546   -78   -18      -6            -2
#> # … with 1 more variable: fiscal_year <dbl>