Compare two columns within a data frame

Compare two columns within a data frame

Compare two conditions within a data frame

compare_columns(df, cols = everything(), calc = list(mean = mean))

compare_conditions(df, x, y, cols = everything(), calc = list(mean = mean))

Arguments

df
data frame

cols
columns to use in comparison

calc
named list of the functions to use, ex: list(mean = mean, sd = sd) 'purrr' style phrases are also supported like list(mean = ~mean(.x, na.rm = TRUE), sd = sd)

x
condition for comparison, same criteria you would use in 'dplyr::filter', used in contrast to the reference group 'y'

y
condition for comparison, same criteria you would use in 'dplyr::filter', used in contrast to the reference group 'x'

Examples

# by default the function will compute the means
compare_columns(flights_jfk, c("dep_delay", "arr_delay"))

#> $mean_arr_delay
#> [1] 0.1459121
#> 
#> $mean_dep_delay
#> [1] 10.41128
#> 

# you can name the output using 'list()' and select the columns using
# any of the 'select()' helpers
compare_columns(flights_jfk, ends_with("delay"), list(avg = mean))

#> $avg_arr_delay
#> [1] 0.1459121
#> 
#> $avg_dep_delay
#> [1] 10.41128
#> 

# compare_conditions() produces a list that can be passed to headline_list()
flights_jfk %>%
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    dep_delay
  )

#> $mean_dep_delay_x
#> [1] 14.03664
#> 
#> $mean_dep_delay_y
#> [1] 2.556334
#> 

.Last.value %>% headline_list()

#> Error in l[[2]]: subscript out of bounds

 # you can return multiple objects to compare
 # 'view_List()' is a helper to see list objects in a compact way
flights_jfk %>%
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    cols = c(dep_delay, arr_delay),
    calc = list(mean = mean, sd = sd)
  ) %>%
  view_list()

#>                     VALUES
#> mean_arr_delay_x  2.621043
#> mean_arr_delay_y -5.216871
#> mean_dep_delay_x 14.036643
#> mean_dep_delay_y  2.556334
#> sd_arr_delay_x   49.396265
#> sd_arr_delay_y   28.515699
#> sd_dep_delay_x   42.733077
#> sd_dep_delay_y   21.753064

# if you want to compare to the average use 'complete.cases(.)'
flights_jfk %>%
  compare_conditions(
    x = (hour > 12),
    y = complete.cases(.),
    dep_delay
  ) %>%
  view_list()

#>                    VALUES
#> mean_dep_delay_x 14.03664
#> mean_dep_delay_y 10.41128