Compare two conditions within a data frame

Compare two conditions within a data frame

compare_conditions(df, x, y, .cols = everything(), .fns = lst(mean), ...)

Arguments

df

data frame

x

condition for comparison, same criteria you would use in 'dplyr::filter', used in contrast to the reference group 'y'

y

condition for comparison, same criteria you would use in 'dplyr::filter', used in contrast to the reference group 'x'

.cols

columns to use in comparison

.fns

named list of the functions to use, ex: list(avg = mean, sd = sd) 'purrr' style phrases are also supported like list(mean = ~mean(.x, na.rm = TRUE), sd = sd) and dplyr::lst(mean, sd) will create a list(mean = mean, sd = sd)

Examples

# compare_conditions works similar to dplyr::across()
flights_jfk |>
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    .cols = dep_delay
  )
#> # A tibble: 1 × 2
#>   mean_dep_delay_x mean_dep_delay_y
#>              <dbl>            <dbl>
#> 1             14.0             2.56

# because data frames are just fancy lists, you pass the result to headline_list()
flights_jfk |>
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    .cols = dep_delay
  ) |>
  headline_list("a difference of {delta} minutes")
#> [1] "a difference of 11.5 minutes"

 # you can return multiple objects to compare
 # 'view_List()' is a helper to see list objects in a compact way
flights_jfk |>
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    .cols = c(wind_dir, visib),
    .fns = dplyr::lst(mean, sd)
  ) |>
  view_list()
#>                     VALUES
#> mean_visib_x      9.619185
#> mean_visib_y      9.484816
#> mean_wind_dir_x 205.104405
#> mean_wind_dir_y 212.942207
#> sd_visib_x        1.400228
#> sd_visib_y        1.699849
#> sd_wind_dir_x    91.027026
#> sd_wind_dir_y   102.533374

# if you want to compare x to the overall average, use y = TRUE
flights_jfk |>
  compare_conditions(
    x = (hour > 12),
    y = TRUE,
    .cols = dep_delay
  )
#> # A tibble: 1 × 2
#>   mean_dep_delay_x mean_dep_delay_y
#>              <dbl>            <dbl>
#> 1             14.0             10.4