Compose phrases that describe differences in the data

Compose phrases that describe differences in the data

headline(...)

# S3 method for default
headline(
  compare,
  reference,
  headline = "{trend} of {delta} ({orig_values})",
  ...,
  if_match = "There was no difference.",
  trend_phrases = headliner::trend_terms(),
  plural_phrases = NULL,
  orig_values = "{c} vs. {r}",
  n_decimal = 1,
  round_all = TRUE,
  multiplier = 1,
  return_data = FALSE
)

# S3 method for list
headline(x, compare, reference, ...)

Arguments

...
Arguments passed on to headline.default

compare
numeric value to compare against reference (base) value

reference
numeric value that 'compare' value will be compared against

headline
a string to format the final output. Uses glue syntax

if_match
string to display if numbers match, uses glue syntax

trend_phrases
list of values to use for when y is more than x, y is the same as x, or y is less than x. You can pass it just trend_terms (the default) and call the result with "...{trend}..." or pass is a named list (see examples)

plural_phrases
named list of values to use when difference (delta) is singular (delta = 1) or plural (delta != 1)

orig_values
a string to display the two original values. Uses glue syntax. {c} = the 'compare' value, and {r} = 'reference'

n_decimal
numeric value to limit the number of decimal places in the returned values.

round_all
logical value to indicate if all values should be rounded. When FALSE, the values will return with no modification. When TRUE (default) all values will be round to the length specified by 'n_decimal'.

multiplier
number indicating the scaling factor. When multiplier = 1 (default), 0.25 will return 0.25. When multiplier = 100, 0.25 will return 25.

return_data
logical to indicate whether function should return the phrase components used to compose the headline

x
a list with values to compare, if named, can call by name

See also

Examples

# values can be manually entered, some headlines are provided by default
headline(10, 8)

#> increase of 2 (10 vs. 8)

headline(8, 10)

#> decrease of 2 (8 vs. 10)

headline(10, 10)

#> There was no difference.

# most likely you'll edit the headline by hand
headline(
  compare = 10,
  reference = 8,
  headline = "There was a ${delta} {trend} vs last year"
)

#> There was a $2 increase vs last year

# you can also adjust the phrasing of higher/lower values
headline(
  compare = 10,
  reference = 8,
  headline = "Group A was {trend} by {delta_p}%.",
  trend_phrases = trend_terms(more = "higher", less = "lower")
 )

#> Group A was higher by 25%.

# a phrase about the comparion can be edited by providing glue syntax
# 'c' = the 'compare' value, 'r' = 'reference'
headline(10, 8, orig_values = "{c} to {r} people")

#> increase of 2 (10 to 8 people)

# you can also add phrases for when the difference = 1 or not
headline(
  compare = 10,
  reference = 8,
  plural_phrases = list(
    were = plural_phrasing(single = "was", multi = "were"),
    people = plural_phrasing(single = "person", multi = "people")
  ),
  headline = "there {were} {delta} {people}"
)

#> there were 2 people

# you can also adjust the rounding, although the default is 1
headline(0.1234, 0.4321)

#> decrease of 0.3 (0.1 vs. 0.4)

headline(0.1234, 0.4321, n_decimal = 3)

#> decrease of 0.309 (0.123 vs. 0.432)

# or use a multiplier
headline(0.1234, 0.4321, multiplier = 100)

#> decrease of 30.9 (12.3 vs. 43.2)

# there are many components you can assemble
headline(
  compare = 16,
  reference = 8,
  headline = "there was {article_delta_p}% {trend}, \\
  {add_article(trend)} of {delta} ({orig_values})"
)

#> there was a 100% increase, an increase of 8 (16 vs. 8)

# compare_conditions() produces a list that can be passed to headline()
 mtcars %>%
   compare_conditions(
     compare = cyl == 4,
     reference = cyl == 6,
     cols = c(mpg)
   ) %>%
   headline(
     headline =
       "4-cylinder cars get an average of {delta} {trend} miles \\
       per gallon than 6-cylinder cars ({orig_values}).",
     trend_phrases = trend_terms("more", "less")
   )

#> 4-cylinder cars get an average of 6.9 more miles per gallon than 6-cylinder cars (26.7 vs. 19.7).