Compose phrases that describe differences in the data

Compose phrases that describe differences in the data

headline(
  x,
  y,
  headline = "{trend} of {delta} ({orig_values})",
  ...,
  if_match = "There was no difference.",
  trend_phrases = headliner::trend_terms(),
  plural_phrases = NULL,
  orig_values = "{x} vs. {y}",
  n_decimal = 1,
  round_all = TRUE,
  multiplier = 1,
  return_data = FALSE
)

headline_list(
  l,
  headline = "{trend} of {delta} ({orig_values})",
  x,
  y,
  ...,
  if_match = "There was no difference.",
  trend_phrases = headliner::trend_terms(),
  plural_phrases = NULL,
  orig_values = "{x} vs. {y}",
  n_decimal = 1,
  round_all = TRUE,
  multiplier = 1,
  return_data = FALSE
)

Arguments

x

a numeric value to compare to the reference value of 'y'

y

a numeric value to act as a control for the 'x' value

headline

a string to format the final output. Uses glue syntax

...

arguments passed to glue_data

if_match

string to display if numbers match, uses glue syntax

trend_phrases

list of values to use for when y is more than x, y is the same as x, or y is less than x. You can pass it just trend_terms (the default) and call the result with "...{trend}..." or pass is a named list (see examples)

plural_phrases

named list of values to use when difference (delta) is singular (delta = 1) or plural (delta != 1)

orig_values

a string using glue syntax. example: ({x} vs {y})

n_decimal

numeric value to limit the number of decimal places in the returned values.

round_all

logical value to indicate if all values should be rounded. When FALSE, the values will return with no modification. When TRUE (default) all values will be round to the length specified by 'n_decimal'.

multiplier

number indicating the scaling factor. When multiplier = 1 (default), 0.25 will return 0.25. When multiplier = 100, 0.25 will return 25.

return_data

logical to indicate whether function should return the phrase components used to compose the headline

l

a list with values to compare, if named, can call by name

See also

Examples

# values can be manually entered, some headlines are provided by default
headline(10, 8)
#> [1] "increase of 2 (10 vs. 8)"
headline(8, 10)
#> [1] "decrease of 2 (8 vs. 10)"
headline(1:3, 3:1)
#> [1] "decrease of 2 (1 vs. 3)"  "There was no difference."
#> [3] "increase of 2 (3 vs. 1)" 

# most likely you'll edit the headline by hand
headline(
  x = 10,
  y = 8,
  headline = "There was a ${delta} {trend} vs last year"
)
#> [1] "There was a $2 increase vs last year"

# you can also adjust the phrasing of higher/lower values
headline(
  x = 10,
  y = 8,
  headline = "Group A was {trend} by {delta_p}%.",
  trend_phrases = trend_terms(more = "higher", less = "lower")
 )
#> [1] "Group A was higher by 25%."

# a phrase about the comparion can be edited by providing glue syntax
# 'c' = the 'compare' value, 'r' = 'reference'
headline(10, 8, orig_values = "{x} to {y} people")
#> [1] "increase of 2 (10 to 8 people)"

# you can also add phrases for when the difference = 1 or not
headline(
  x = 10,
  y = 8,
  plural_phrases = list(
    were = plural_phrasing(single = "was", multi = "were"),
    people = plural_phrasing(single = "person", multi = "people")
  ),
  headline = "there {were} {delta} {people}"
)
#> [1] "there were 2 people"

# you can also adjust the rounding, the default is 1
headline(0.1234, 0.4321)
#> [1] "decrease of 0.3 (0.1 vs. 0.4)"
headline(0.1234, 0.4321, n_decimal = 3)
#> [1] "decrease of 0.309 (0.123 vs. 0.432)"
# or use a multiplier
headline(0.1234, 0.4321, multiplier = 100)
#> [1] "decrease of 30.9 (12.3 vs. 43.2)"

# there are many components you can assemble
headline(
  x = 16,
  y = 8,
  headline = "there was {article_delta_p}% {trend}, \\
  {add_article(trend)} of {delta} ({orig_values})"
)
#> [1] "there was a 100% increase, an increase of 8 (16 vs. 8)"

# compare_conditions() and compare_columns() produce lists that can be
# passed to headline_list()
flights_jfk %>%
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    dep_delay
  ) %>%
  headline_list()
#> [1] "increase of 11.5 (14 vs. 2.6)"

# if you have more than 2 list items, you can specify them by name
list(
  x = 1,
  y = 2,
  z = 3
 ) %>%
  headline_list(
    x = x,
    y = z
  )
#> [1] "decrease of 2 (1 vs. 3)"